Allow JDK XML alias field naming
